-u favicon

+++ src/home/dashboard/gestion-de-archivos/components/buttons/buttonDelete.tsx
@@ -3,6 +3,7 @@ import { IFileSelected } from '../../../../../common/interface';
 import { useDeleteFilesMutation, useGetFilesMutation } from '../../../../../store/api/filesApi';
 import { useDispatch } from 'react-redux';
 import { deleteFilesSelected, updateFiles } from '../../../../../store/slice/file_managerSlice';
+import { confirmAction } from '../../../../../common/helpers';
 
 export const ButtonDeleteFile = () => {
 	const dispatch = useDispatch();
@@ -15,10 +16,12 @@ export const ButtonDeleteFile = () => {
 		dispatch(updateFiles(data.data));
 	};
 	const handle = async () => {
-		const { data }: any = await deleteFiles({ data: filesSelected.map((file: IFileSelected) => file.uuid) });
-		if (data.status == 200) {
-			handleGetData();
-		}
+		confirmAction('¿Seguro que deseas eliminar los archivos seleccionados?', async () => {
+			const { data }: any = await deleteFiles({ data: filesSelected.map((file: IFileSelected) => file.uuid) });
+			if (data.status == 200) {
+				handleGetData();
+			}
+		});
 	};
 	return (
 		<div
